backend/requirements.txt
Django
djangorestframework
celery
django-filter
django-rest-auth
requests
python-baseconv
django-phonenumber-field
phonenumbers
Pillow
django-allauth
django-extensions
gunicorn
psycopg2-binary
django-crispy-forms
pdf2image
selenium
sklearn
PyPDF2
bs4
fpdf
pyqrcode